initial_material_ui step closer to grid and drawer
getting grid and the material-ui drawer to play nicely together is
proving to be a challenge. In this commit we have the header and drawer
integrated, but the content for the drawer is the only part moving out
of the way. The grid does not appear to be obey the drawer