apollo_state about halfway through this process with limited success
getting the WOOT screen. Now we need to use apollo state to determine
which screen to display in our SPA. At the moment, it checks for the
existence of a cookie, but since we are using localStorage and there
should be a copy in Apollo cache there is no need for this approach.