apollo_state tests working except for client token async issue
token async is a problem because localStorage is a sync lib and handling
it async has not yet been sorted out for the client. It is complicated
by the fact that the apollo client initialization with apollo boost is
designed to handle all async data, but getting the token is a triple
layer of abstraction while `{}` style strings in ES6 can handle only 2
that I have tried (JSON and getToken). If we can find a way to get the
token key's value in the same statement it might work